+++ horovod/ray/__init__.py
@@ -1,4 +1,5 @@
-from .runner import RayExecutor, BaseHorovodWorker
+from .worker import BaseHorovodWorker
+from .runner import RayExecutor
 from .elastic import ElasticRayExecutor
 
 __all__ = ["RayExecutor", "BaseHorovodWorker", "ElasticRayExecutor"]
